Reduce code duplication in analyze
