+++ src/main/java/org/polypheny/simpleclient/scenario/PolyphenyScenario.java
@@ -1,0 +1,65 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2019-2025 The Polypheny Project
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a
+ * copy of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+package org.polypheny.simpleclient.scenario;
+
+import java.io.File;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
+import java.util.Properties;
+import java.util.concurrent.ConcurrentHashMap;
+import lombok.extern.slf4j.Slf4j;
+import org.polypheny.simpleclient.QueryMode;
+import org.polypheny.simpleclient.executor.JdbcExecutor;
+
+@Slf4j
+public abstract class PolyphenyScenario extends Scenario {
+
+    protected long executeRuntime;
+    protected final Map<Integer, String> queryTypes;
+    protected final List<Long> measuredTimes = Collections.synchronizedList( new LinkedList<>() );
+    protected final Map<Integer, List<Long>> measuredTimePerQueryType = new ConcurrentHashMap<>();
+
+
+    public PolyphenyScenario( JdbcExecutor.ExecutorFactory executorFactory, boolean commitAfterEveryQuery, boolean dumpQueryList, QueryMode queryMode ) {
+        super( executorFactory, commitAfterEveryQuery, dumpQueryList, queryMode );
+        queryTypes = new HashMap<>();
+    }
+
+
+    @Override
+    public void analyze( Properties properties, File outputDirectory ) {
+        properties.put( "measuredTime", calculateMean( measuredTimes ) );
+
+        measuredTimePerQueryType.forEach( ( templateId, time ) -> calculateResults( queryTypes, properties, templateId, time ) );
+        properties.put( "queryTypes_maxId", queryTypes.size() );
+        properties.put( "executeRuntime", executeRuntime / 1000000000.0 );
+        properties.put( "numberOfQueries", measuredTimes.size() );
+        properties.put( "throughput", measuredTimes.size() / (executeRuntime / 1000000000.0) );
+    }
+
+}
