+++ src/main/java/org/polypheny/simpleclient/scenario/graph/GraphBench.java
@@ -31,7 +31,6 @@ import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
-import java.util.Properties;
 import java.util.Random;
 import java.util.Vector;
 import java.util.concurrent.ConcurrentHashMap;
@@ -51,7 +50,7 @@ import org.polypheny.simpleclient.query.QueryBuilder;
 import org.polypheny.simpleclient.query.QueryListEntry;
 import org.polypheny.simpleclient.scenario.EvaluationThread;
 import org.polypheny.simpleclient.scenario.EvaluationThreadMonitor;
-import org.polypheny.simpleclient.scenario.Scenario;
+import org.polypheny.simpleclient.scenario.PolyphenyScenario;
 import org.polypheny.simpleclient.scenario.graph.queryBuilder.CountNodePropertyBuilder;
 import org.polypheny.simpleclient.scenario.graph.queryBuilder.CreateGraphDatabase;
 import org.polypheny.simpleclient.scenario.graph.queryBuilder.DeleteNodeBuilder;
@@ -66,7 +65,7 @@ import org.polypheny.simpleclient.scenario.graph.queryBuilder.UnwindBuilder;
 
 
 @Slf4j
-public class GraphBench extends Scenario {
+public class GraphBench extends PolyphenyScenario {
 
     public static final String GRAPH_NAMESPACE = "test";
     public static boolean EXPECTED_RESULT = true;
@@ -273,17 +272,6 @@ public class GraphBench extends Scenario {
                 throw new RuntimeException( "Unexpected interrupt", e );
             }
         }
-    }
-
-
-    @Override
-    public void analyze( Properties properties, File outputDirectory ) {
-        properties.put( "measuredTime", calculateMean( measuredTimes ) );
-        measuredTimePerQueryType.forEach( ( templateId, time ) -> calculateResults( queryTypes, properties, templateId, time ) );
-        properties.put( "queryTypes_maxId", queryTypes.size() );
-        properties.put( "executeRuntime", executeRuntime / 1000000000.0 );
-        properties.put( "numberOfQueries", measuredTimes.size() );
-        properties.put( "throughput", measuredTimes.size() / (executeRuntime / 1000000000.0) );
     }
